pyproject.toml
[project]
name = "calorie-app"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.68.0",
    "gradio>=6.9.0",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
]
